split JWT / encryption related tests out remove babel from tests
